reorganize tests files
