--- test/flutter_love_provider_test.dart
@@ -1,7 +1,0 @@
-import 'tests/dispatch_x_test.dart' as dispatch_x_test;
-import 'tests/system_providers_test.dart' as system_providers_test;
-
-void main() {
-  dispatch_x_test.main();
-  system_providers_test.main();
-}
